add TemplatedLRInstruction type with template-driven keyframe animation
Introduces a new atomic instruction type `templated_lr` that generates
animation segments from named templates containing per-role keyframes
with relative positions and facings. Also relaxes BeatsSchema from
integer-only to allow fractional beat values.

+++ src/contraCore.ts
@@ -4,7 +4,7 @@ import { z } from "zod";
 import { NORTH, SOUTH } from "./geometry";
 import type { AssertExtends } from "./utils";
 
-export const BeatsSchema = z.number().int();
+export const BeatsSchema = z.number();
 export type Beats = z.infer<typeof BeatsSchema>;
 
 export const HandSchema = z.enum(["left", "right"]);
